+++ packages/snaps-controllers/src/interface/index.ts
@@ -1,1 +1,19 @@
-export * from './SnapInterfaceController';
+export type {
+  ExpandedView,
+  SnapInterfaceControllerGetStateAction,
+  SnapInterfaceControllerState,
+  SnapInterfaceControllerStateChangeEvent,
+  StoredInterface,
+} from './SnapInterfaceController';
+export { SnapInterfaceController } from './SnapInterfaceController';
+export type {
+  SnapInterfaceControllerCreateInterfaceAction,
+  SnapInterfaceControllerDeleteInterfaceAction,
+  SnapInterfaceControllerGetInterfaceAction,
+  SnapInterfaceControllerGetInterfaceStateAction,
+  SnapInterfaceControllerMethodActions,
+  SnapInterfaceControllerResolveInterfaceAction,
+  SnapInterfaceControllerSetInterfaceDisplayedAction,
+  SnapInterfaceControllerUpdateInterfaceAction,
+  SnapInterfaceControllerUpdateInterfaceStateAction,
+} from './SnapInterfaceController-method-action-types';